Patch libtwh5.so permenantly
The tofwerk DLL `libtwh5.so` doesn't have a well specified path for its
dependency `libtwtool.so`. Currently, in production, the tofwerk DLL is
patched using patchelf in the backend Dockerfile at build-time.

However, this doesn't resolve the problem for development environments
in Linux. Additionally, this adds a repeated step at build time which
requires an additionally dependency, slowing down our builds.

This uses the following `patchelf` command on `libtwh5.so` that tells
it to look for the dependency in its own folder:

  patchelf libtwh5.so --add-rpath '$ORIGIN'

For provenance, the a copy of the original, unpatched DLL is preserved.

As a result of these changes, the extra steps in the build can be
removed from the Dockerfile.