Make the acquisition tab default & move tab logic
Since no batch is focused when the app is first loaded, the batch tab
is empty by default. In order to provide a more sane default, this
sets the acquisitions tab to be the default.

Additionally the tab auto-switching logic is moved the Dashboard
component to the tab store, and so minor bugs are resolved.