feat: auto switch to and from the batch tab
This implements changes that automatically switch to the batch
tab when opening a non-empty batch. Conversely if the batch tab is open
and all batches are deselected or an empty batch is opened, we switch
back to the default tab.

In order to achieve this, lifecycle hooks are added to the frontend
store modules:
 - onRefocus - triggered after a refocus reloads the module
 - onEvent - triggered after an event reloads the module

The implementation is nuanced, since we needed to address the following
edge cases:
 - When deleting the last sample of a batch, we need to switch away
   from the batch tab
 - When deleting a sample which is not the last of the batch, we
   must not switch to or from the batch tab

The implementation results in the behavior that when processing or
importing samples into a batch, the batch tab is not automatically
opened. This was deemed either neutral or even desirable, but was
mostly accepted since an implementation which does the opposite
would be hard to reconcile with the above edgecases.

Resolves #776