fix: right batch overview chart size on tab switch
This ensures that when alternating tabs and split resizes, the batch
overview chart is always sized correctly. The solution was to force
charts to fully rerender completely when switching tabs by using the
PrimeVue <Tabs> component's `lazy` prop.

This incedently also fixes the issue for other charts, e.g. the
spectrum chart.

Resolves #775