refactor: match params and visualized API
We mistakingly reinitialize UI parameter values to defaults when
saving match params. This exposes a new argument to the
app.data.match.param store's init action to allow skipping the reset
step and ensure that we reset the initial values while not
reseting the current values.

Digging deeper, we uncover some rough edges in the match store API.
The match params and match visualized stores need to be carefully
orchastrated together to ensure consistent state.

This refactors to stores so that they interactions are internalized
in store actions. The result is a safer and simpler API that aim to
help maintain consistent state without call patterns.

Resolves #763
Resolves #783
Resolves #788