fix: extend label length batch overview chart
Visibility of long labels on the x-axis of the batch overview chart
is currently not good. This partially mitigates the issue by:

 1. Increasing the height of the chart
 2. Increasing the minreducedheight property
 3. Turning tick labels to 45 degrees

The combination makes the chart's actual size more variable, since it
maximum label length can influnce it. However, long labels are more
readible, and batches with shorter labels are visualized with a larger
chart.

Resolves #647