Ensure mascope_hardware init in mascope_lib
A recent PR that adds an API to load signal directly from h5 files
included commit 9922ea69 which adds `mascope_hardware` as a dependency
of `mascope_lib`. However, the hardware runtime was not initialized,
resulting in the following error when running the file converter:

  Attribute Error: 'NoneType' object has no attribute 'config'

This ensures the `mascope_lib` runtime module initializes the
`mascope_hardware` runtime and resolves the error.