if_exists arg for compute_all_sample_file_peaks
In order to prepare for features recomputing peaks after instrument
config updates, this propagates the `if_exist` argument from the
`compute_all_sample_file_peaks` function to the `detect_peaks`
function. The `append` value is used as the default to maintain
previous behavior, but the argument allows passing `replace` in order
to recompute the peaks from scratch.