Match parameter sidebar
The match parameter user interface was long considered a bit awkward.
Popover elements needed to be launched seperately for different param
groups as well as the menu for saving parameters.

This moves all parameters and the menu into a unified sidebar. The
sidebar obscures the target and sample browsers, and does not use a
modal overlay. The result is an improved UX for managing match params.