Stacktrace printing option for logging infra
Debugging has historically been impeded by the fact that our callstack
is deep, and yet our log messages do not have a way of reflecting that
stack.

This adds a pretty stacktrace printing feature leveraging the rich
Traceback class used in our logger already for printing Exception
tracebacks.

To use this, add the argument `trace=True` to a logger call:

  runtime.logger.info(
    "this message will come with a stacktrace",
    trace=True
  )