module.js: ensure multiselect focus consistency
Multiselectable data modules can still use the `focus` method to ensure
only one record is selected. Currently, if the record provided is not
actually loaded to the module, selection will become `[ undefined ]`,
i.e. an array with undefined in it. This is not a consistent state.

This ensures that using the `focus` method in a multiselectable data
module never results in a dirty state.