Refactor: sample browser stores
Context menu logic in the sample stores has been a source of confusion
and bugs, as well as inconsistent UX. In order to facilitate mitigation
of these issues, this refactors key logic in the sample browser
components into stores:

 - batchContext for the batch table's context menu
 - sampleContext for the sample table's context menu
 - customizerPopover for the sample table customizer popover
 - clipboard for working with copy / paste operations

A key change is that context menus now use the right clicked row (a.k.a
the context row) to launch dialogs and other actions. Previously focus
was the prime trigger of actions.

Resolves #773