Help store performance fix
Previously the help store use a vueuse helper called useMouseInElement
to identify whether the mouse is in a help element in order to decide
when to show the help card.

The helper provides detailed reactive information, including mouse
position. this resulting in excessive updates to the store with
information we don't need for the feature.

This excludes this information, taking only the `isOutside` reactive
variable from `useMouseInElement`. The result removes a massive amount
of unnecessary state updates from the frontend.