Seperate sample context menu component
Initial right clicks on sample rows in the sample browsers have
often not successfully opened the context menu. Subsequent clicks have
then worked as expected.

After careful debugging, it was discovered that this is due to the
context menu component unmounting every time a batch is closed.

This seperates a SampleContextMenu component from the SampleTable
component and imports it directly into PaneBrowserSample, ensuring the
context menu is no mounted with batch expansion, but rather with the
browser parent component.