Refactor: runtime seperation of concerns
Lots of confusion has surrounded the runtime library which underpins
the orchastration of Mascope's various modules. While some essential
complexity is necessarily required to address this challenge, some of
the confusion was due to an awkward implementation.

This attempts to improve the situation by refactoring the runtime
library into a set of classes that explicitly expose interfaces to
key concepts in the runtime library:

  - Runtime - the main runtime instance class, used in each module
    and exposing both global and local (module specific) functionality
  - RuntimeModule - the local runtime module interface, including its
    name and configuration.
  - RuntimeEnv - the runtime environment, including its name, path and
    a method for resolving paths relative to it.
  - RuntimeConfigLoader - a helper to facilitate loading and resolving
    the runtime's configuration
  - RuntimeLoggerLoader - a helper to facilitate configuring the logger
    for the runtime

This helps encapsulate key functionality into contained classes, and
provides a clear parent-child relationship with the main `Runtime` class
at the root. Each child class has access to the parent through a `_root`
attribute, with type hints only being imported when type checking in
order to prevent circular import errors.

This also centralizes configuration defaults across all environments:
`base.mascope.toml` is moved from the default runtime to the runtime
library, and configuration loading is modified to always read it from
that path.

Resolves #825