Runtime secret reading method
Secret reading is currently done with add-hoc code, and uses hardcoded
fallbacks when running in dev mode.

This adds a method to the runtime library to help load secrets into
the application. The method accepts two parameters:
 - envvar - the environment variable for resolving path in prod
 - path - the path relative to the secrets folder for resolving in dev

The method then tries the following in sequence
 - try the path provided by the envvar
 - try path argument, relative to $MASCOPE_PATH/secrets
 - try path argument, relative to $MASCOPE_PATH/runtime/secrets

It returns the first file it finds. If no file is found, an error is
raised.

Resolves #803