Add download data listening to the frontend
This allows any background job to automatically trigger a file download
on the frontend. The endpoint needs to add a `download` field to its
`_notification_data`. The frontend will then trigger a download from
the /api/temp endpoint, using the value provided in the download field
as the path parameter and filename.