Autonomous Python install script
The setup scripts used to install, reinstall and uninstall mascope
have a flaw: they depend in part of the command line itself. Any
problems with the command line cause the script to bug out.

This adds an autonomous Python script with no depedencies but the
standard library, which can install and uninstall Mascope's CLI
and modules.

This can be used in setup scripts to manage Mascope's installation
without dependence on the CLI.

The windows.ps1, ubuntu.sh and nixos.sh scripts are updated to use
the autonomous python script to install Mascope. Other dependencies
(like pipx, poetry and docker) are installed using the system's
native package manager.

This allows us to remove the un/install commands from CLI, which are
now deprecated.