Refactor: migrate to uv & reorganize monorepo (#853)
* Remove .legacy folder

* Refactor: migrate to uv & reorganize monorepo

Migrate to uv from Poetry, improving development experience.
- Organize project as a uv 'workspace' that shares a single
  virtual environment.
- The root project in Mascope's case is just a container for
  the venv, which is why `src/mascope` is empty (besides a
  single __init__.py).
- Run `npm install` on `mascope dev run`; combined with
  uv's autosync feature, it means we no longer need to
  run setup scripts to reinstall when switching branches.
- Therefore, mascope.py can be deprecated and many configuration
  options can be removed.

Entry points:
- Scripts are streamline to reduce duplication, with entrypoints
  for db ops and service launch refined using Typer.
- Scripts are brought to the root-level pyproject.toml to provide
  a unified overview.
- Entrypoints are also exposed through the main CLI for convenience
  and discoverability.

Libraries refactor:
- Split `mascope_hardware` into
  * `mascope_tofwerk` for the TOF API
  * `mascope_thermo` for the
- Split `mascope_lib` into
  * `mascope_file` for sample io, name & record management (no deps)
  * `mascope_signal` for signal processing & depdent on hardware libs
  * `mascope_chem` for chemical calculations
This helps manage which depedency goes where, and in particular
reduces the TOF Agent's dependencies.

Resolves #828
Resolves #826

CLI: mascope log command

Add a `log query` command to enable querying log files to the CLI.
The runtime loggin API is first extended with the functionality, which
is then exposed to the CLI.

Also add a `mascope logs gc` command to delete old logs, based on
specified --retain interval or --before datetime.

The functionality leverages DuckDB to glob logfiles into memory
and then query them. A variety of query options are exposed both
to the runtime and CLI APIs:

 - time range with --from, --to and --interval options
 - grepping with --grep and --grep-ctx options (the latter determining
  how many rows before and after a match to show).
 - runtime mode with --prod and --dev (defaulting to --dev).
 - limit the total number of lines printed with --max

The logs are printed with loguru, approximately configured the
same as when used in the `mascope dev run` and `mascope prod run`
commands.

Why this implementation? I spend a day or so searching on and off for
off-the-shelf solutions for reading JSONL log files. I originally
selected this format because I thought it was easy to find a good
tool; it turns out this format is standard in the cloud, and most
solutions did not suit self-hosted projects like Mascope. Additionally,
many solutions did not support Windows.

The implementation need to rapidly injest many JSON files and parse
them, and then query the data. This is not something that is easy to
implement efficiently in Python. Luckily, DuckDB does it efficiently
out of the box.

For printing, it seemed natural to try and reuse our existing logger.

Resolves #507

* Fix File Agent bugs

* Fix README typo

* Remove legacy code from thermo library

* Update README

Add mentioning about activating the virtual environment

* Fix filestore path bug

---------

Co-authored-by: Oskari Kausiala <oskari.kausiala@karsa.fi>