Add Peak browser & peak assignment interface
This adds a peak browser to the spectrum tab, allowing users to browse
peaks by focusing them. The spectrum chart responds by highlight the
region of the spectrum with the peak.

Additionally, the new cheminfo endpoint is used to preview potential
matches by searching for the selected peak's m/z in the cheminfo
API and then aggregating matches for the returned formulas.

Finally, the user can add new compounds to the currently focused
target collection.

Resolves #616
Resolves #468