Add sample file download API
This adds a background task and route for downloading sample files.
A list of sample item IDs is provided and the task then conditionally
zips an archive if mutliple files are selected.

To enable this, we add support for `null` download responses in the
frontend, as well as supporting warning download notifications.

In the backend, we also modify the temp file route to add a filename
specification. This will make FastAPI add it to Content-Disposition,
which can resolve certain issues with downloads.