Make delete_sample_file controller remove files
Currently the delete_sample_file controller only deletes the database
record corresponding to the sample file. The actual folder of processed
sample file data in the filestore is not removed.

This modifies the controller to delete the files using shutils.rmtree.
Additionally a `acquisitions_reload` event is emitted, to be used
by the frontend to reload acquisition data when sample files are
deleted.