feat(cli): add `env sync` command
This adds a `mascope env sync` command that uses rsync to allow
backing, migrating and synchronizing runtime envs locally or remotely.

Destination synlinked directories are maintained to allow mounted
filestore drives in production environments not to be broken by the
sync process. A flag to skip the filestore is provided to allow for
faster transfers.