fix(frontend): load newly created workspace
Loading of a newly created workspace was plagued by a race condition:
logic in module.js tries to refocus the previously focused workspace
when data is reloaded, and a watcher in the workspace dialog attempts
to focus the newly created workspace.

This creates a new API in module.js to lazily focus a record the first
time a load makes it available, resolving the bug.