feat(agent): Create Mascope CSV Export Agent
Add new agent to facilitate Mascope match data exporting into text files in semi real time.

The application periodically checks for new samples in a specified workspace, and computes matches for configurable set of target compounds for each new sample, and writes the results into a text file.

This is per request of UHEL people for the CLOUD campaign.