fix(ui): Use peak height instead of intensity for assignment preview
Fixes a bug where "sample_peak_intensity" was used directly as the reference peak height for assignment preview. For the TOF, peak area is used as the intensity, which caused it to be wrongly visualized. Now, it fetches the corresponding peak from the peak store, and always uses the height instead.