feat(lib): Improve resolution function fitting for TOF
Improves the fitting of the TOF resolution function by implementing a logarithmic outlier filtering method for FWHM values.

Adds iterative Huber weighting to the linear fit of FWHM for improved coefficient estimation and employs adaptive adjustments to ensure a reasonable dynamic range of the rational polynomial. This avoids premature plateauing in the resolution function.