fix(ui): Fix splitter panel inner component sizing issue
Fixes a bug where in absence of "mascope-dashboard-split" and "mascope-browser-split" in the local storage, components inside the splitter panels are sized wrong. This was due to conflicting default sizes in "split.js" and "Dashboard.vue".

Also, changes the default horizontal split from 20/80 to 40/60 due to recent changes in the target browser, requiring more space.